refactor: rename to spaghetti-comb-v2.nvim
