+++ lua/spaghetti-comb-v2/tests/init.lua
@@ -1,4 +1,4 @@
--- Test runner for nvim-navigation-breadcrumbs
+-- Test runner for spaghetti-comb-v2
 local M = {}
 
 -- Run all tests
@@ -7,10 +7,10 @@ function M.run_all()
   
   -- Collect all test sets
   local test_sets = {
-    require('nvim-navigation-breadcrumbs.tests.history_spec'),
-    require('nvim-navigation-breadcrumbs.tests.ui_spec'),
-    require('nvim-navigation-breadcrumbs.tests.navigation_spec'),
-    require('nvim-navigation-breadcrumbs.tests.integration_spec'),
+    require('spaghetti-comb-v2.tests.history_spec'),
+    require('spaghetti-comb-v2.tests.ui_spec'),
+    require('spaghetti-comb-v2.tests.navigation_spec'),
+    require('spaghetti-comb-v2.tests.integration_spec'),
   }
   
   -- Run tests
@@ -22,25 +22,25 @@ end
 -- Run specific test category
 function M.run_history()
   local MiniTest = require('mini.test')
-  local test_set = require('nvim-navigation-breadcrumbs.tests.history_spec')
+  local test_set = require('spaghetti-comb-v2.tests.history_spec')
   MiniTest.run(test_set)
 end
 
 function M.run_ui()
   local MiniTest = require('mini.test')
-  local test_set = require('nvim-navigation-breadcrumbs.tests.ui_spec')
+  local test_set = require('spaghetti-comb-v2.tests.ui_spec')
   MiniTest.run(test_set)
 end
 
 function M.run_navigation()
   local MiniTest = require('mini.test')
-  local test_set = require('nvim-navigation-breadcrumbs.tests.navigation_spec')
+  local test_set = require('spaghetti-comb-v2.tests.navigation_spec')
   MiniTest.run(test_set)
 end
 
 function M.run_integration()
   local MiniTest = require('mini.test')
-  local test_set = require('nvim-navigation-breadcrumbs.tests.integration_spec')
+  local test_set = require('spaghetti-comb-v2.tests.integration_spec')
   MiniTest.run(test_set)
 end
